website/requirements.txt
amply==0.1.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
Django==3.1.2
django-wkhtmltopdf==3.3.0
docutils==0.16
et-xmlfile==1.0.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.5
PuLP==2.3.1
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
